+++ src/demo/index.js
@@ -5,5 +5,6 @@ import App from './AppMinimumUsage';
 // import App from './AppAllProps';
 // import App from './AppWithUmountCamera';
 // import App from './AppWithDynamicProperties';
+// import App from './AppSaveToLocalFile';
 
 ReactDOM.render(<App />, document.getElementById('root'));